+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. This change
 - added a `kindly-compute` endpoint - for requesting server computations - with @RamNarayan-27
 - added `kind/scittle`
 - added automatic recognition of `kind/scittle` and `kind/reagent` when nested inside `kind/hiccup`
+- added basic support for a Scittle header script with initial setup
 
 ## [2-beta27] - 2024-12-24
 - added a `:browse` option (default `true`) to determine whether to open a browser tab when the server is started